requirements.txt
matplotlib==3.7.2
numpy==1.24.3
opencv_python==4.9.0.80
Pillow==9.5.0
Pillow==10.2.0
scikit_image==0.19.3
scipy==1.10.1
segment_anything==1.0
torch==2.1.0
tqdm==4.66.1
